Add UVM infrastructure for PPU program tests
This changeset introduces the barebone nescesasry components
for running test programs on the processor with UVM. Note
that this infrastructure cannot be used to verify the processor
itself. So far the structure is largely modelled after the
omnibus memory tests. The hx_omnibus_programtest_sequence
loads and runs a series of program test cases on the ppu.

This commit contains also some other miscellaneous changes:

Makefile.ncsim was extended in the following way:

- because of a stupid limitation of bash or glibc (it only
  supports 127 command line arguments) we had to
  introduce a argument list for irun

- timing checks had to be disabled because otherwise
  the sram macros produced garbage results.

- verbosity of irun / make was reduced

- UVM_RELEASE_NOTES were supressed

Bus_widener had a bug in the response path that was fixed.

"Automatic" testdiscovery for programtests:

- It is now relatively easy to add an arbitrary
  number of new testcases to be run

- Each directory with testcases should contain a
  testcase file which stores a number of parameter
  of the test. Currently this is format:

  <name : string> <should_read_back : bool> <timeout : int>

  Depending on the paramters set the test is executed and
  then conditionally a readback of the mailbox is performed,
  or the test is terminated and a error raised if it does
  not complete within the timeout

- programtest_sequence is now "smarter" it only writes
  the test binary to the memory and does not initialize
  the rest of the memory. It also only reads back conditionally
  a much smaller part of the memory (the mailbox) on test
  completion or none at all. Finally it optionally
  times out as before. Overall this drastically reduces testcase
  execution time, because useless work is not done anymore.

- fixed typo in omnibus_address_pkg.sv

TODO:

While this can now be considered a somewhat suitable test environment
very few meaningful tests do actually successfully execute at the
moment

Change-Id: I0be6ad911eb20bd5eb5448a2bbb03233d601085a