Fix ppu omnibus hierachy
Partially WIP, will be somewhat extendend. The processor uses
byte addressing but the omnibus operates on words of a certain size,
with an optional byte enable. Therefore in particular the
ppu only ever generates 29..0 non zero bits plus a byte enable and
this results in the unconnected omnibus_master if the select bit
is set to 31. Similar reason for external instruction access.

Change-Id: I545dc21cc51a75cdbd034d90a2e84a909fd78d2f