Add additional (functional) processor tests
These tests have been adapted from github.com/riscv-tests/
(hence the LICENSE) to powerpc, the ppc_test.h is a
minimalistic test shim for our current power processor.

Unfortunately lwz.S actually encounters a bug in the
omnibus hierachy, this has to be investigated in the next
patch.

This changeset adds additional tests to the processor and
implements a simulation backend for the l2 to instruction memory
queues.

Change-Id: Ia3fe4fb6c76ccb93cb413d897d56a9f7331920ef