Further vector tests + support for top_pads environment
This changeset adds first functional tests of the vector unit
(comparison with expected result is done in 32bit words) and
some fixes to the clocks in the ppu module. It also adapts
the tests in hx_chip_control so that they can work with the
hx_top_pads test environment.

Change-Id: I9e9cfb18f6e918603000d0dcb2d85bd2855c3e47